format & remove old add op test
